+++ HISTORY.md
@@ -1,8 +1,11 @@
 # History
 
+## 0.10.0 (Unreleased)
+* Fix per-host rate-limiting for Redis and Postgres backends
+* If both `per_host=True` and a `bucket_name` is specified, use `bucket_name` as a bucket prefix
+
 ## 0.9.3 (2026-04-02)
-* Fix compatibility with `RedisBucket`
-* Fix compatibility with `PostgresBucket`
+* Fix bucket initialization for `RedisBucket` and `PostgresBucket`
 * Use built-in support for pickling `Limiter` from pyrate-limiter 4.1.0
 
 ## 0.9.2 (2026-02-27)